feat(nexus): added command handler response
