+++ crates/nexus/src/aggregate/aggregate_root.rs
@@ -1,0 +1,101 @@
+use super::{
+    Aggregate, AggregateCommandHandler, AggregateState, AggregateType,
+    command_handler::CommandHandlerResponse,
+};
+use crate::Command;
+
+/// Concrete implementation managing an aggregate's state based on events.
+/// Generic over the aggregate definition provided by `AT: AggregateType`.
+#[derive(Debug)]
+pub struct AggregateRoot<AT: AggregateType> {
+    id: AT::Id,
+    state: AT::State,
+    /// The version loaded from the event store (number of events historically applied).
+    version: u64,
+    /// Events generated by the last command execution, not yet persisted.
+    uncommitted_events: Vec<AT::Event>,
+}
+
+impl<AT> Aggregate for AggregateRoot<AT>
+where
+    AT: AggregateType,
+{
+    type Id = AT::Id;
+    type Event = AT::Event;
+    type State = AT::State;
+
+    fn id(&self) -> &Self::Id {
+        &self.id
+    }
+
+    fn state(&self) -> &Self::State {
+        &self.state
+    }
+
+    fn version(&self) -> u64 {
+        self.version
+    }
+
+    fn take_uncommitted_events(&mut self) -> Vec<Self::Event> {
+        std::mem::take(&mut self.uncommitted_events)
+    }
+}
+
+impl<AT> AggregateRoot<AT>
+where
+    AT: AggregateType,
+{
+    /// Creates a new aggregate instance with default state and version 0.
+    pub fn new(id: AT::Id) -> Self {
+        Self {
+            id,
+            state: AT::State::default(),
+            version: 0,
+            uncommitted_events: Vec::new(),
+        }
+    }
+
+    /// Rehydrates aggregate state by applying historical events.
+    pub fn load_from_history(id: AT::Id, history: impl IntoIterator<Item = AT::Event>) -> Self {
+        let mut state = AT::State::default();
+        let mut version = 0u64;
+
+        for event in history {
+            state.apply(&event);
+            version += 1;
+        }
+
+        Self {
+            id,
+            state,
+            version,
+            uncommitted_events: Vec::new(),
+        }
+    }
+
+    /// Gets the effective current version, including uncommitted events.
+    /// Useful for repositories implementing optimistic concurrency checks.
+    pub fn current_version(&self) -> u64 {
+        self.version + self.uncommitted_events.len() as u64
+    }
+
+    pub async fn execute<C, Handler, Services>(
+        &mut self,
+        command: C,
+        handler: &Handler,
+        services: &Services,
+    ) -> Result<C::Result, C::Error>
+    where
+        C: Command,
+        Handler: AggregateCommandHandler<C, Services, State = AT::State>,
+        Services: Send + Sync + ?Sized,
+    {
+        let CommandHandlerResponse { events, result } =
+            handler.handle(&self.state, command, services).await?;
+        for event in &events {
+            self.state.apply(event);
+        }
+        self.uncommitted_events.extend(events);
+        Ok(result)
+    }
+}
