+++ crates/nexus/src/aggregate/mod.rs
@@ -1,5 +1,11 @@
-use super::{Command, DomainEvent};
-use std::{boxed::Box, fmt::Debug, hash::Hash, pin::Pin};
+use crate::DomainEvent;
+use std::{fmt::Debug, hash::Hash};
+
+pub mod aggregate_root;
+pub mod command_handler;
+
+pub use aggregate_root::AggregateRoot;
+pub use command_handler::AggregateCommandHandler;
 
 /// Represents the actual state data of an aggregate.
 /// It's rebuilt by applying events and must know how to apply them.
@@ -27,7 +33,7 @@ pub trait AggregateType: Send + Sync + Debug + Copy + Clone + 'static {
 
 /// Provides a standard interface for accessing aggregate properties.
 /// Implemented by `AggregateRoot`.
-pub trait Aggregate: Send + Sync + 'static {
+pub trait Aggregate: Debug + Send + Sync + 'static {
     type Id: Send + Sync + Debug + Eq + Hash + Clone + 'static;
     type Event: DomainEvent + PartialEq;
     type State: AggregateState<Event = Self::Event>;
@@ -39,131 +45,6 @@ pub trait Aggregate: Send + Sync + 'static {
     /// Takes ownership of newly generated events for saving.
     fn take_uncommitted_events(&mut self) -> Vec<Self::Event>;
 }
-
-pub type CommandHandlerResult<State, C> =
-    Result<Vec<<State as AggregateState>::Event>, <C as Command>::Error>;
-
-pub type CommandHandlerFuture<'a, State, C> =
-    Pin<Box<dyn Future<Output = CommandHandlerResult<State, C>> + Send + 'a>>;
-
-/// Encapsulates the pure domain logic for a specific command/state combination.
-/// Implement this for each command that modifies an aggregate.
-pub trait AggregateCommandHandler<State, C, Services>: Send + Sync
-where
-    State: AggregateState,
-    C: Command,
-    Services: Send + Sync + ?Sized,
-{
-    /// Processes the command against the current state using optional services.
-    /// Returns a list of events if successful, or the command-specific domain error.
-    fn handle<'a>(
-        &'a self,
-        state: &'a State,
-        command: C,
-        services: &'a Services,
-    ) -> CommandHandlerFuture<'a, State, C>;
-
-    /// Determines the result (`C::Result`) to return upon successful handling.
-    fn derive_result(&self) -> C::Result;
-}
-
-/// Concrete implementation managing an aggregate's state based on events.
-/// Generic over the aggregate definition provided by `AT: AggregateType`.
-#[derive(Debug)]
-pub struct AggregateRoot<AT: AggregateType> {
-    id: AT::Id,
-    state: AT::State,
-    /// The version loaded from the event store (number of events historically applied).
-    version: u64,
-    /// Events generated by the last command execution, not yet persisted.
-    uncommitted_events: Vec<AT::Event>,
-}
-
-impl<AT> Aggregate for AggregateRoot<AT>
-where
-    AT: AggregateType,
-{
-    type Id = AT::Id;
-    type Event = AT::Event;
-    type State = AT::State;
-
-    fn id(&self) -> &Self::Id {
-        &self.id
-    }
-
-    fn state(&self) -> &Self::State {
-        &self.state
-    }
-
-    fn version(&self) -> u64 {
-        self.version
-    }
-
-    fn take_uncommitted_events(&mut self) -> Vec<Self::Event> {
-        std::mem::take(&mut self.uncommitted_events)
-    }
-}
-
-impl<AT> AggregateRoot<AT>
-where
-    AT: AggregateType,
-{
-    /// Creates a new aggregate instance with default state and version 0.
-    pub fn new(id: AT::Id) -> Self {
-        Self {
-            id,
-            state: AT::State::default(),
-            version: 0,
-            uncommitted_events: Vec::new(),
-        }
-    }
-
-    /// Rehydrates aggregate state by applying historical events.
-    pub fn load_from_history(id: AT::Id, history: impl IntoIterator<Item = AT::Event>) -> Self {
-        let mut state = AT::State::default();
-        let mut version = 0u64;
-
-        for event in history {
-            state.apply(&event);
-            version += 1;
-        }
-
-        Self {
-            id,
-            state,
-            version,
-            uncommitted_events: Vec::new(),
-        }
-    }
-
-    /// Gets the effective current version, including uncommitted events.
-    /// Useful for repositories implementing optimistic concurrency checks.
-    pub fn current_version(&self) -> u64 {
-        self.version + self.uncommitted_events.len() as u64
-    }
-
-    pub async fn execute<C, Handler, Services>(
-        &mut self,
-        command: C,
-        handler: &Handler,
-        services: &Services,
-    ) -> Result<C::Result, C::Error>
-    where
-        C: Command,
-        Handler: AggregateCommandHandler<AT::State, C, Services>,
-        Services: Send + Sync + ?Sized,
-    {
-        let generated_events = handler.handle(&self.state, command, services).await?;
-        for event in &generated_events {
-            self.state.apply(event);
-        }
-        self.uncommitted_events.extend(generated_events);
-        let result = handler.derive_result();
-        Ok(result)
-    }
-}
-
-// // --- Example Definitions (User Aggregate) ---
 
 // // Marker type for User Aggregate
 // #[derive(Debug, Clone, Copy)]
